fix(ios): inline pipbridgeplugin into appdelegate so xcode can find it
Loose .swift files in ios/Runner aren't auto-discovered by Xcode — they
have to be explicitly added to Runner.xcodeproj/project.pbxproj.  The
v0.0.297 release pipeline failed at 'Build iOS IPA' with:

    Cannot find 'PipBridgePlugin' in scope

because PipBridgePlugin.swift was created on disk but never registered
in the project file.  Move the class + its delegate sibling into
AppDelegate.swift (which is in the project) and delete the orphan
file.  Same code, no semantic change; just stops fighting the Xcode
build system.

Tracks the v0.0.297 publish failure from #839's release run.
